Sepereate member request from JCDB_query
