+++ ESD/nbproject/private/private.xml
@@ -2,6 +2,9 @@
 <project-private xmlns="http://www.netbeans.org/ns/project-private/1">
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group/>
+        <group>
+            <file>file:/C:/Users/Tim%20Lam/Documents/NetBeansProjects/Enterprise-Project/ESD/src/java/model/JdbcQry.java</file>
+            <file>file:/C:/Users/Tim%20Lam/Documents/NetBeansProjects/Enterprise-Project/ESD/src/java/model/member_request.java</file>
+        </group>
     </open-files>
 </project-private>
