+++ ESD/src/java/Controller/login.java
@@ -5,6 +5,7 @@
  */
 package Controller;
 
+import com.sun.corba.se.pept.transport.Connection;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.sql.SQLException;
@@ -15,7 +16,7 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import model.JdbcQry;
+import model.member_request;
 import java.sql.*;
 /**
  *
@@ -40,7 +41,8 @@ public class login extends HttpServlet {
         String password=request.getParameter("password");
         
          
-        JdbcQry j = new JdbcQry((Connection) request.getServletContext().getAttribute("connection"));
+        member_request j = new member_request();
+       
         boolean b = j.validateLogin(username, password);
         if (b==true) {
             boolean a = j.checkAdmin(username);
