+++ ESD/src/java/Controller/register.java
@@ -14,6 +14,7 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import model.JdbcQry;
+import model.member_request;
 import model.Member;
 import java.util.*;
 import java.text.ParseException;
@@ -40,7 +41,7 @@ public class register extends HttpServlet {
         Member m = new Member();
         response.setContentType("text/html;charset=UTF-8");
         String username = request.getParameter("username");
-        JdbcQry j = new JdbcQry((Connection) request.getServletContext().getAttribute("connection"));        
+        member_request j = new member_request();        
         String name = request.getParameter("name");
         String password = request.getParameter("password");
         String address = request.getParameter("address");
@@ -54,7 +55,6 @@ public class register extends HttpServlet {
             m.setDor(sql);
 
         } catch (ParseException e) {
-            e.printStackTrace();
         }
         
         m.setId(username);
